rookyard

+++ src/io/iomapserialize.cpp
@@ -9,13 +9,12 @@
 
 #include "pch.hpp"
 
-#include "io/iomapserialize.hpp"
-#include "io/iologindata.hpp"
-#include "game/game.hpp"
-#include "items/bed.hpp"
+#include "io/iomapserialize.h"
+#include "game/game.h"
+#include "items/bed.h"
 
 void IOMapSerialize::loadHouseItems(Map* map) {
-	Benchmark bm_context;
+	int64_t start = OTSYS_TIME();
 
 	DBResult_ptr result = Database::getInstance().storeQuery("SELECT `data` FROM `tile_store`");
 	if (!result) {
@@ -35,7 +34,7 @@ void IOMapSerialize::loadHouseItems(Map* map) {
 			continue;
 		}
 
-		std::shared_ptr<Tile> tile = map->getTile(x, y, z);
+		Tile* tile = map->getTile(x, y, z);
 		if (!tile) {
 			continue;
 		}
@@ -46,36 +45,22 @@ void IOMapSerialize::loadHouseItems(Map* map) {
 		}
 
 		while (item_count--) {
-			if (auto houseTile = std::dynamic_pointer_cast<HouseTile>(tile)) {
-				const auto house = houseTile->getHouse();
-				if (house->getOwner() == 0) {
-					g_logger().trace("Skipping load item from house id: {}, position: {}, house does not have owner", house->getId(), house->getEntryPosition().toString());
-					house->clearHouseInfo(false);
-					continue;
-				}
-			}
-
-			loadItem(propStream, tile, true);
+			loadItem(propStream, tile);
 		}
 	} while (result->next());
-	g_logger().info("Loaded house items in {} milliseconds", bm_context.duration());
+	SPDLOG_INFO("Loaded house items in {} seconds", (OTSYS_TIME() - start) / (1000.));
 }
 
 bool IOMapSerialize::saveHouseItems() {
-	bool success = DBTransaction::executeWithinTransaction([]() {
-		return SaveHouseItemsGuard();
-	});
-
-	if (!success) {
-		g_logger().error("[{}] Error occurred saving houses", __FUNCTION__);
-	}
-
-	return success;
-}
-
-bool IOMapSerialize::SaveHouseItemsGuard() {
+	int64_t start = OTSYS_TIME();
 	Database &db = Database::getInstance();
 	std::ostringstream query;
+
+	// Start the transaction
+	DBTransaction transaction;
+	if (!transaction.begin()) {
+		return false;
+	}
 
 	// clear old tile data
 	if (!db.executeQuery("DELETE FROM `tile_store`")) {
@@ -87,7 +72,7 @@ bool IOMapSerialize::SaveHouseItemsGuard() {
 	PropWriteStream stream;
 	for (const auto &[key, house] : g_game().map.houses.getHouses()) {
 		// save house items
-		for (const auto &tile : house->getTiles()) {
+		for (HouseTile* tile : house->getTiles()) {
 			saveTile(stream, tile);
 
 			size_t attributesSize;
@@ -106,13 +91,16 @@ bool IOMapSerialize::SaveHouseItemsGuard() {
 		return false;
 	}
 
-	return true;
+	// End the transaction
+	bool success = transaction.commit();
+	SPDLOG_INFO("Saved house items in {} seconds", (OTSYS_TIME() - start) / (1000.));
+	return success;
 }
 
-bool IOMapSerialize::loadContainer(PropStream &propStream, std::shared_ptr<Container> container) {
+bool IOMapSerialize::loadContainer(PropStream &propStream, Container* container) {
 	while (container->serializationCount > 0) {
 		if (!loadItem(propStream, container)) {
-			g_logger().warn("Deserialization error for container item: {}", container->getID());
+			SPDLOG_WARN("Deserialization error for container item: {}", container->getID());
 			return false;
 		}
 		container->serializationCount--;
@@ -120,56 +108,49 @@ bool IOMapSerialize::loadContainer(PropStream &propStream, std::shared_ptr<Conta
 
 	uint8_t endAttr;
 	if (!propStream.read<uint8_t>(endAttr) || endAttr != 0) {
-		g_logger().warn("Deserialization error for container item: {}", container->getID());
+		SPDLOG_WARN("Deserialization error for container item: {}", container->getID());
 		return false;
 	}
 	return true;
 }
 
-bool IOMapSerialize::loadItem(PropStream &propStream, std::shared_ptr<Cylinder> parent, bool isHouseItem /*= false*/) {
+bool IOMapSerialize::loadItem(PropStream &propStream, Cylinder* parent) {
 	uint16_t id;
 	if (!propStream.read<uint16_t>(id)) {
 		return false;
 	}
 
-	std::shared_ptr<Tile> tile = nullptr;
+	Tile* tile = nullptr;
 	if (parent->getParent() == nullptr) {
 		tile = parent->getTile();
 	}
 
 	const ItemType &iType = Item::items[id];
-	if (iType.isBed() || iType.movable || !tile || iType.isCarpet()) {
+	if (iType.moveable || !tile || iType.isCarpet()) {
 		// create a new item
-		auto item = Item::CreateItem(id);
+		Item* item = Item::CreateItem(id);
 		if (item) {
 			if (item->unserializeAttr(propStream)) {
-				// Remove only not movable and not sleeper bed
-				auto bed = item->getBed();
-				if (isHouseItem && iType.isBed() && bed && bed->getSleeper() == 0 && !iType.movable) {
-					return false;
-				}
-				std::shared_ptr<Container> container = item->getContainer();
+				Container* container = item->getContainer();
 				if (container && !loadContainer(propStream, container)) {
+					delete item;
 					return false;
 				}
 
 				parent->internalAddThing(item);
 				item->startDecaying();
 			} else {
-				g_logger().warn("Deserialization error in {}", id);
-
+				SPDLOG_WARN("Deserialization error in {}", id);
+				delete item;
 				return false;
 			}
 		}
 	} else {
 		// Stationary items like doors/beds/blackboards/bookcases
-		std::shared_ptr<Item> item = nullptr;
+		Item* item = nullptr;
 		if (const TileItemVector* items = tile->getItemList()) {
-			for (auto &findItem : *items) {
+			for (Item* findItem : *items) {
 				if (findItem->getID() == id) {
-					item = findItem;
-					break;
-				} else if (iType.m_transformOnUse && findItem->getID() == iType.m_transformOnUse) {
 					item = findItem;
 					break;
 				} else if (iType.isDoor() && findItem->getDoor()) {
@@ -184,26 +165,26 @@ bool IOMapSerialize::loadItem(PropStream &propStream, std::shared_ptr<Cylinder> 
 
 		if (item) {
 			if (item->unserializeAttr(propStream)) {
-				std::shared_ptr<Container> container = item->getContainer();
+				Container* container = item->getContainer();
 				if (container && !loadContainer(propStream, container)) {
 					return false;
 				}
 
 				g_game().transformItem(item, id);
 			} else {
-				g_logger().warn("Deserialization error in {}", id);
+				SPDLOG_WARN("Deserialization error in {}", id);
 			}
 		} else {
 			// The map changed since the last save, just read the attributes
-			auto dummy = Item::CreateItem(id);
+			std::unique_ptr<Item> dummy(Item::CreateItem(id));
 			if (dummy) {
 				dummy->unserializeAttr(propStream);
-				std::shared_ptr<Container> container = dummy->getContainer();
+				Container* container = dummy->getContainer();
 				if (container) {
 					if (!loadContainer(propStream, container)) {
 						return false;
 					}
-				} else if (std::shared_ptr<BedItem> bedItem = std::dynamic_pointer_cast<BedItem>(dummy)) {
+				} else if (BedItem* bedItem = dynamic_cast<BedItem*>(dummy.get())) {
 					uint32_t sleeperGUID = bedItem->getSleeper();
 					if (sleeperGUID != 0) {
 						g_game().removeBedSleeper(sleeperGUID);
@@ -215,8 +196,8 @@ bool IOMapSerialize::loadItem(PropStream &propStream, std::shared_ptr<Cylinder> 
 	return true;
 }
 
-void IOMapSerialize::saveItem(PropWriteStream &stream, std::shared_ptr<Item> item) {
-	std::shared_ptr<Container> container = item->getContainer();
+void IOMapSerialize::saveItem(PropWriteStream &stream, const Item* item) {
+	const Container* container = item->getContainer();
 
 	// Write ID & props
 	stream.write<uint16_t>(item->getID());
@@ -234,21 +215,19 @@ void IOMapSerialize::saveItem(PropWriteStream &stream, std::shared_ptr<Item> ite
 	stream.write<uint8_t>(0x00); // attr end
 }
 
-void IOMapSerialize::saveTile(PropWriteStream &stream, std::shared_ptr<Tile> tile) {
+void IOMapSerialize::saveTile(PropWriteStream &stream, const Tile* tile) {
 	const TileItemVector* tileItems = tile->getItemList();
 	if (!tileItems) {
 		return;
 	}
 
-	std::forward_list<std::shared_ptr<Item>> items;
+	std::forward_list<Item*> items;
 	uint16_t count = 0;
-	for (auto &item : *tileItems) {
-		if (item->getID() == ITEM_BATHTUB_FILLED_NOTMOVABLE) {
-			std::shared_ptr<Item> tub = Item::CreateItem(ITEM_BATHTUB_FILLED);
-			items.push_front(tub);
-			++count;
-			continue;
-		} else if (!item->isSavedToHouses()) {
+	for (Item* item : *tileItems) {
+		const ItemType &it = Item::items[item->getID()];
+
+		// Note that these are NEGATED, ie. these are the items that will be saved.
+		if (!(it.moveable || it.isCarpet() || item->getDoor() || (item->getContainer() && !item->getContainer()->empty()) || it.canWriteText || item->getBed())) {
 			continue;
 		}
 
@@ -263,7 +242,7 @@ void IOMapSerialize::saveTile(PropWriteStream &stream, std::shared_ptr<Tile> til
 		stream.write<uint8_t>(tilePosition.z);
 
 		stream.write<uint32_t>(count);
-		for (std::shared_ptr<Item> item : items) {
+		for (const Item* item : items) {
 			saveItem(stream, item);
 		}
 	}
@@ -272,31 +251,15 @@ void IOMapSerialize::saveTile(PropWriteStream &stream, std::shared_ptr<Tile> til
 bool IOMapSerialize::loadHouseInfo() {
 	Database &db = Database::getInstance();
 
-	DBResult_ptr result = db.storeQuery("SELECT `id`, `owner`, `new_owner`, `paid`, `warnings` FROM `houses`");
+	DBResult_ptr result = db.storeQuery("SELECT `id`, `owner`, `paid`, `warnings` FROM `houses`");
 	if (!result) {
 		return false;
 	}
 
 	do {
-		auto houseId = result->getNumber<uint32_t>("id");
-		const auto house = g_game().map.houses.getHouse(houseId);
+		House* house = g_game().map.houses.getHouse(result->getNumber<uint32_t>("id"));
 		if (house) {
-			uint32_t owner = result->getNumber<uint32_t>("owner");
-			int32_t newOwner = result->getNumber<int32_t>("new_owner");
-			// Transfer house owner
-			auto isTransferOnRestart = g_configManager().getBoolean(TOGGLE_HOUSE_TRANSFER_ON_SERVER_RESTART, __FUNCTION__);
-			if (isTransferOnRestart && newOwner >= 0) {
-				g_game().setTransferPlayerHouseItems(houseId, owner);
-				if (newOwner == 0) {
-					g_logger().debug("Removing house id '{}' owner", houseId);
-					house->setOwner(0);
-				} else {
-					g_logger().debug("Setting house id '{}' owner to player GUID '{}'", houseId, newOwner);
-					house->setOwner(newOwner);
-				}
-			} else {
-				house->setOwner(owner, false);
-			}
+			house->setOwner(result->getNumber<uint32_t>("owner"), false);
 			house->setPaidUntil(result->getNumber<time_t>("paid"));
 			house->setPayRentWarnings(result->getNumber<uint32_t>("warnings"));
 		}
@@ -305,7 +268,7 @@ bool IOMapSerialize::loadHouseInfo() {
 	result = db.storeQuery("SELECT `house_id`, `listid`, `list` FROM `house_lists`");
 	if (result) {
 		do {
-			const auto &house = g_game().map.houses.getHouse(result->getNumber<uint32_t>("house_id"));
+			House* house = g_game().map.houses.getHouse(result->getNumber<uint32_t>("house_id"));
 			if (house) {
 				house->setAccessList(result->getNumber<uint32_t>("listid"), result->getString("list"));
 			}
@@ -315,45 +278,40 @@ bool IOMapSerialize::loadHouseInfo() {
 }
 
 bool IOMapSerialize::saveHouseInfo() {
-	bool success = DBTransaction::executeWithinTransaction([]() {
-		return SaveHouseInfoGuard();
-	});
-
-	if (!success) {
-		g_logger().error("[{}] Error occurred saving houses info", __FUNCTION__);
-	}
-
-	return success;
-}
-
-bool IOMapSerialize::SaveHouseInfoGuard() {
 	Database &db = Database::getInstance();
 
-	std::ostringstream query;
-	DBInsert houseUpdate("INSERT INTO `houses` (`id`, `owner`, `paid`, `warnings`, `name`, `town_id`, `rent`, `size`, `beds`) VALUES ");
-	houseUpdate.upsert({ "owner", "paid", "warnings", "name", "town_id", "rent", "size", "beds" });
-
-	for (const auto &[key, house] : g_game().map.houses.getHouses()) {
-		std::string values = fmt::format("{},{},{},{},{},{},{},{},{}", house->getId(), house->getOwner(), house->getPaidUntil(), house->getPayRentWarnings(), db.escapeString(house->getName()), house->getTownId(), house->getRent(), house->getSize(), house->getBedCount());
-
-		if (!houseUpdate.addRow(values)) {
-			return false;
-		}
-	}
-
-	if (!houseUpdate.execute()) {
+	DBTransaction transaction;
+	if (!transaction.begin()) {
 		return false;
 	}
 
-	DBInsert listUpdate("INSERT INTO `house_lists` (`house_id` , `listid` , `list`, `version`) VALUES ");
-	listUpdate.upsert({ "list", "version" });
-	auto version = getTimeUsNow();
+	if (!db.executeQuery("DELETE FROM `house_lists`")) {
+		return false;
+	}
+
+	std::ostringstream query;
+	for (const auto &[key, house] : g_game().map.houses.getHouses()) {
+		query << "SELECT `id` FROM `houses` WHERE `id` = " << house->getId();
+		DBResult_ptr result = db.storeQuery(query.str());
+		if (result) {
+			query.str(std::string());
+			query << "UPDATE `houses` SET `owner` = " << house->getOwner() << ", `paid` = " << house->getPaidUntil() << ", `warnings` = " << house->getPayRentWarnings() << ", `name` = " << db.escapeString(house->getName()) << ", `town_id` = " << house->getTownId() << ", `rent` = " << house->getRent() << ", `size` = " << house->getTiles().size() << ", `beds` = " << house->getBedCount() << " WHERE `id` = " << house->getId();
+		} else {
+			query.str(std::string());
+			query << "INSERT INTO `houses` (`id`, `owner`, `paid`, `warnings`, `name`, `town_id`, `rent`, `size`, `beds`) VALUES (" << house->getId() << ',' << house->getOwner() << ',' << house->getPaidUntil() << ',' << house->getPayRentWarnings() << ',' << db.escapeString(house->getName()) << ',' << house->getTownId() << ',' << house->getRent() << ',' << house->getTiles().size() << ',' << house->getBedCount() << ')';
+		}
+
+		db.executeQuery(query.str());
+		query.str(std::string());
+	}
+
+	DBInsert stmt("INSERT INTO `house_lists` (`house_id` , `listid` , `list`) VALUES ");
 
 	for (const auto &[key, house] : g_game().map.houses.getHouses()) {
 		std::string listText;
 		if (house->getAccessList(GUEST_LIST, listText) && !listText.empty()) {
-			query << house->getId() << ',' << GUEST_LIST << ',' << db.escapeString(listText) << ',' << version;
-			if (!listUpdate.addRow(query)) {
+			query << house->getId() << ',' << GUEST_LIST << ',' << db.escapeString(listText);
+			if (!stmt.addRow(query)) {
 				return false;
 			}
 
@@ -361,18 +319,18 @@ bool IOMapSerialize::SaveHouseInfoGuard() {
 		}
 
 		if (house->getAccessList(SUBOWNER_LIST, listText) && !listText.empty()) {
-			query << house->getId() << ',' << SUBOWNER_LIST << ',' << db.escapeString(listText) << ',' << version;
-			if (!listUpdate.addRow(query)) {
+			query << house->getId() << ',' << SUBOWNER_LIST << ',' << db.escapeString(listText);
+			if (!stmt.addRow(query)) {
 				return false;
 			}
 
 			listText.clear();
 		}
 
-		for (std::shared_ptr<Door> door : house->getDoors()) {
+		for (Door* door : house->getDoors()) {
 			if (door->getAccessList(listText) && !listText.empty()) {
-				query << house->getId() << ',' << door->getDoorId() << ',' << db.escapeString(listText) << ',' << version;
-				if (!listUpdate.addRow(query)) {
+				query << house->getId() << ',' << door->getDoorId() << ',' << db.escapeString(listText);
+				if (!stmt.addRow(query)) {
 					return false;
 				}
 
@@ -381,13 +339,9 @@ bool IOMapSerialize::SaveHouseInfoGuard() {
 		}
 	}
 
-	if (!listUpdate.execute()) {
+	if (!stmt.execute()) {
 		return false;
 	}
 
-	if (!db.executeQuery(fmt::format("DELETE FROM `house_lists` WHERE `version` < {}", version))) {
-		return false;
-	}
-
-	return true;
+	return transaction.commit();
 }
